feat: Add new SVG icons for the application
This commit adds two new SVG icons, "Interrogacion.svg" and "Logo.svg", to the assets/icon directory. These icons will be used in the application's user interface. The "Interrogacion.svg" icon is used to display an information message, while the "Logo.svg" icon represents the logo of the application.

The "Interrogacion.svg" icon is a 16x16 SVG image with a blue fill color. It is used in the text-info__instructions section of the index.html file to indicate that only lowercase letters without accents are allowed.

The "Logo.svg" icon is a 32x48 SVG image with a dark blue fill color. It is used in the header section of the index.html file to display the logo of the application.

These new icons enhance the visual appeal of the application and provide a better user experience.