chore: Update CSS to hide message container on smaller screens
This commit modifies the index.css file to hide the .message-container element on smaller screens. The .message-container element was previously displayed, but it is now set to "display: none" in order to improve the user interface on smaller devices.

This change enhances the responsiveness of the application and ensures a better user experience on different screen sizes.